Vec2I collection type optimisations
Implement Vec2IMap on top of LinkedHashMap instead of TreeMap.
Provide a Vec2ISet implemented on top of LinkedHashSet.

These provide O(1) operations as opposed to O(log(n)) operations of
TreeMap/TreeSet. In practice, with the procedural generation algorithms in
use in Metazelda, Zosoko, and other Lenna's Inception algorithms, turn out to
be _significantly_ faster.